+++ src/types.ts
@@ -61,7 +61,7 @@ export type AuthsignalOptions = {
   baseUrl?: string;
   tenantId: string;
   onTokenExpired?: () => void;
-  mode: "rest" | "websocket";
+  mode?: "rest" | "websocket";
 };
 
 export enum AuthsignalWindowMessage {
